refactor validator and its tests
