+++ packages/amplify-graphql-validate-transformer/src/graphql-validate-transformer.ts
@@ -1,95 +1,13 @@
 import { ValidateDirective } from '@aws-amplify/graphql-directives';
-import { TransformerPluginBase, DirectiveWrapper, InvalidDirectiveError } from '@aws-amplify/graphql-transformer-core';
+import { TransformerPluginBase, DirectiveWrapper } from '@aws-amplify/graphql-transformer-core';
 import {
   TransformerContextProvider,
   TransformerSchemaVisitStepContextProvider,
   TransformerPluginProvider,
 } from '@aws-amplify/graphql-transformer-interfaces';
-import {
-  ArgumentNode,
-  DirectiveNode,
-  FieldDefinitionNode,
-  InterfaceTypeDefinitionNode,
-  ObjectTypeDefinitionNode,
-  StringValueNode,
-} from 'graphql';
-import { getBaseType } from 'graphql-transformer-common';
-import { ValidateArguments, ValidateDirectiveConfiguration, ValidationType } from './types';
-
-/**
- * Validates that length validation values (minLength, maxLength) are valid non-negative numbers.
- */
-const validateLengthValue = (config: ValidateDirectiveConfiguration): void => {
-  if (config.type !== 'minLength' && config.type !== 'maxLength') {
-    return;
-  }
-
-  const value = parseFloat(config.value);
-  if (isNaN(value) || !Number.isInteger(value) || value < 0) {
-    throw new InvalidDirectiveError(
-      `${config.type} value must be a positive integer. Received '${config.value}' for field '${config.field.name.value}'`,
-    );
-  }
-};
-
-/**
- * Type guards for validation types
- */
-const isNumericValidation = (type: ValidationType): boolean => {
-  return ['gt', 'lt', 'gte', 'lte'].includes(type);
-};
-
-const isStringValidation = (type: ValidationType): boolean => {
-  return ['minLength', 'maxLength', 'startsWith', 'endsWith', 'matches'].includes(type);
-};
-
-/**
- * Validates that the validation type is compatible with the field type.
- */
-const validateTypeCompatibility = (field: FieldDefinitionNode, validationType: ValidationType): void => {
-  const baseTypeName = getBaseType(field.type);
-
-  if (isNumericValidation(validationType) && baseTypeName !== 'Int' && baseTypeName !== 'Float') {
-    throw new InvalidDirectiveError(
-      `Validation type '${validationType}' can only be used with numeric fields (Int, Float). Field '${field.name.value}' is of type '${baseTypeName}'`,
-    );
-  }
-
-  if (isStringValidation(validationType) && baseTypeName !== 'String') {
-    throw new InvalidDirectiveError(
-      `Validation type '${validationType}' can only be used with String fields. Field '${field.name.value}' is of type '${baseTypeName}'`,
-    );
-  }
-};
-
-/**
- * Validates that there are no duplicate validation types on the same field.
- */
-const validateNoDuplicateTypes = (field: FieldDefinitionNode, currentDirective: DirectiveNode, currentType: ValidationType): void => {
-  for (const peerDirective of field.directives!) {
-    if (peerDirective === currentDirective) {
-      continue;
-    }
-
-    if (peerDirective.name.value === 'validate') {
-      const peerType = peerDirective.arguments?.find((arg: ArgumentNode) => arg.name.value === 'type')?.value as StringValueNode;
-      if (peerType?.value === currentType) {
-        throw new InvalidDirectiveError(
-          `Duplicate @validate directive with type '${currentType}' on field '${field.name.value}'. Each validation type can only be used once per field.`,
-        );
-      }
-    }
-  }
-};
-
-/**
- * Validates all aspects of the @validate directive configuration.
- */
-const validate = (definition: FieldDefinitionNode, directive: DirectiveNode, config: ValidateDirectiveConfiguration): void => {
-  validateTypeCompatibility(definition, config.type as ValidationType);
-  validateNoDuplicateTypes(definition, directive, config.type as ValidationType);
-  validateLengthValue(config);
-};
+import { DirectiveNode, FieldDefinitionNode, InterfaceTypeDefinitionNode, ObjectTypeDefinitionNode } from 'graphql';
+import { ValidateArguments, ValidateDirectiveConfiguration } from './types';
+import { validate } from './validators';
 
 export class ValidateTransformer extends TransformerPluginBase implements TransformerPluginProvider {
   private directiveMap = new Map<string, ValidateDirectiveConfiguration[]>();
